refactor(interfaces): moving components
